Refactor React components and message looping
Refactored the message related components so that MessageList requires Message and Notifications. Message and Notifications make the actual jsx for those specific components while MessageList conditionally calls one or the other, as required, and puts them in a main container.

Added an array.map function to loop through the (currently static) message data and populate the main container with message data. Added keys to each individual message container.

Added a function in componentDidMount which simulates a new message posted, on delay after load.