User input message
The root app now sends a function as a prop to ChatBar. In ChatBar, an event is triggered when the user presses Enter in the message bar, which takes that message and calls the passed function with that data. The called function in the root app then updates the state which rerenders the page with the new message.