Add Change Username Functionality
Added function in app.jsx to change current user in state. Passed that function to ChatBar.jsx. Added an event handler that checks for a username update and passes the new name to the callback function in props.

When a message is submitted, the current user in props is now sent along with the message to ensure that the correct user is assigned to the msg.