Implement Current Users Counter
Added a counter in the server that ++ on new user connection and -- on user close connection. On each of these events, a new message is sent to all current users with the type 'userCountChanged'.

On the front end, this message type is filtered out and the number of current users is changed in state, which triggers a re-render of the new NavBar component.In the NavBar component, the userCount is updated for all online users.